% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypertraps.R
\name{demo.HyperTraPS}
\alias{demo.HyperTraPS}
\title{HyperTraPS demonstration}
\usage{
demo.HyperTraPS()
}
\value{
a ggplot
}
\description{
Test HyperTraPS functionality by performing a fit on synthetic data and reporting a summary
}
