% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypertraps.R
\name{fitted.state.probs}
\alias{fitted.state.probs}
\title{Get state probabilities from a fitted model}
\usage{
\method{fitted}{state.probs}(my.post, prob.set = NA)
}
\arguments{
\item{my.post}{a fitted hypercubic model from HyperTraPS}

\item{prob.set}{a vector of L+1 probabilities giving the probability that 0, 1, ..., L features have been acquired, upon which observation probabilities are then conditioned. Default NA assigns uniform probability to each acquisition number}
}
\value{
a dataframe containing state, observation probability conditional on exactly each state's number of features being acquired, observation probabilities assuming the given probability profiles
}
\description{
Using a fitted model, return a dataframe with observation probabilities of each state, conditional on a given number of features having been acquired
}
