% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getLikelihood}
\alias{getLikelihood}
\title{Get likelihood for an observation, given a fitted model}
\usage{
getLikelihood(
  obs,
  params,
  model,
  initialstate = NULL,
  starttime = NULL,
  endtime = NULL
)
}
\arguments{
\item{obs}{XXX}

\item{params}{XXX}

\item{model}{XXX}

\item{initialstate}{XXX}

\item{starttime}{XXX}

\item{endtime}{XXX}
}
\value{
the likelihood
}
\description{
Get likelihood for an observation, given a fitted model
}
