% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypertraps.R
\name{plotHypercube.bubbles.coarse}
\alias{plotHypercube.bubbles.coarse}
\title{Bubble plot for acquisition ordering, binned version}
\usage{
plotHypercube.bubbles.coarse(
  my.post,
  reorder = FALSE,
  transpose = FALSE,
  bins = 5
)
}
\arguments{
\item{my.post}{a model fit returned from HyperTraPS}

\item{reorder}{logical, whether to reorder features by their mean ordering; default FALSE}

\item{transpose}{logical, whether to transpose the axes of the plot; default FALSE}

\item{bins}{numeric, number of ordering bins to use. Defaults to 5.}
}
\value{
a ggplot object
}
\description{
Creates a plot summarising the ordering of feature acquisitions in a fitted HyperTraPS
model. The circle at ordering i for feature j gives the probability that feature j is
acquired in ordering bin i in an evolutionary process starting with no features and proceeding
to all features. Ordering bins contain sets of specific orderings.
}
\details{
Can be useful to summarise dynamics when there are lots of features.
}
\examples{
observations <- matrix(c(0,0,0,
                         0,0,1,
                         0,1,1,
                         1,1,1), byrow=TRUE, ncol=3)
fitted.cube <- HyperTraPS(observations)
plotHypercube.bubbles.coarse(fitted.cube)
}
