% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypertraps.R
\name{plotHypercube.bubbles}
\alias{plotHypercube.bubbles}
\title{Bubble plot for acquisition ordering}
\usage{
plotHypercube.bubbles(
  my.post,
  reorder = FALSE,
  transpose = FALSE,
  p.color = "#000000"
)
}
\arguments{
\item{my.post}{a model fit returned from HyperTraPS}

\item{reorder}{logical, whether to reorder features by their mean ordering; default FALSE}

\item{transpose}{logical, whether to transpose the axes of the plot; default FALSE}

\item{p.color}{color for the bubbles (default black)}
}
\value{
a ggplot object
}
\description{
Creates a plot summarising the ordering of feature acquisitions in a fitted HyperTraPS
model. The circle at ordering i for feature j gives the probability that feature j is
acquired at order i in an evolutionary process starting with no features and proceeding
to all features. For example, P = 0.2 at i = 3, j = 2 means there is a 0.2 probability that
feature 2 is the third to be acquired.
}
\examples{
observations <- matrix(c(0,0,0,
                         0,0,1,
                         0,1,1,
                         1,1,1), byrow=TRUE, ncol=3)
fitted.cube <- HyperTraPS(observations)
plotHypercube.bubbles(fitted.cube)
}
