% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypertraps.R
\name{plotHypercube.influencegraph}
\alias{plotHypercube.influencegraph}
\title{Visualize pairwise influences (graph)}
\usage{
plotHypercube.influencegraph(
  my.post,
  featurenames = TRUE,
  use.regularised = FALSE,
  use.final = FALSE,
  thresh = 0.05,
  cv.thresh = Inf,
  label.size = 2,
  red.green = FALSE
)
}
\arguments{
\item{my.post}{A model fit returned from HyperTraPS.}

\item{featurenames}{Either logical (TRUE = use names from model fit), or a character vector of feature names. Default TRUE.}

\item{use.regularised}{Logical, whether to use the regularised (by parameter pruning) parameterisation. Defaults to FALSE.}

\item{use.final}{Logical, whether to use only the final parameterisation. Makes sense for point estimates (e.g. from simulated annealing), not for Bayesian posteriors. Defaults to FALSE.}

\item{thresh}{A numeric describing the threshold strength of an interaction to be plotted. Defaults to 0.05}

\item{cv.thresh}{Numeric. Upper threshold of CV for interactions to be plotted. Defaults to Inf.}

\item{label.size}{A numeric describing label size in mm. Defaults to 2}

\item{red.green}{A logical controlling red-green colour palette. Default FALSE.}
}
\value{
a ggplot
}
\description{
plot pairwise influences between features in the L^2 or L^3 picture, as a graph
}
\examples{
observations <- matrix(c(0,0,0,
                         0,0,1,
                         0,1,1,
                         1,1,1), byrow=TRUE, ncol=3)
fitted.cube <- HyperTraPS(observations)
plotHypercube.influencegraph(fitted.cube, featurenames=FALSE)
}
