% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypertraps.R
\name{plotHypercube.lik.trace}
\alias{plotHypercube.lik.trace}
\title{Likelihood trace}
\usage{
plotHypercube.lik.trace(my.post)
}
\arguments{
\item{my.post}{a model fit returned from HyperTraPS}
}
\value{
a ggplot object containing the diagnostic trace
}
\description{
Creates a plot of likelihood throughout an inference run, for diagnostics. The working likelihoods
is the red line; two independent refits are given by black dashed lines. We want the three lines
to overlap (converged likelihood estimate), have constant mean (MCMC run converged), and have no
periods of stasis (MCMC run not frozen).
}
\details{
If these conditions are not met, consider respectively more walkers, longer runs, smaller kernel.
}
\examples{
observations <- matrix(c(0,0,0,
                         0,0,1,
                         0,1,1,
                         1,1,1), byrow=TRUE, ncol=3)
fitted.cube <- HyperTraPS(observations)
plotHypercube.lik.trace(fitted.cube)
}
