% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypertraps.R
\name{plotHypercube.regularisation}
\alias{plotHypercube.regularisation}
\title{Regularisation info and AIC}
\usage{
plotHypercube.regularisation(my.post)
}
\arguments{
\item{my.post}{A model fit returned from HyperTraPS, including the regularisation process}
}
\value{
a ggplot
}
\description{
Creates a plot describing the process of regularisation by parameter pruning.
}
\examples{
observations <- matrix(c(0,0,0,
                         0,0,1,
                         0,1,1,
                         1,1,1), byrow=TRUE, ncol=3)
fitted.cube <- HyperTraPS(observations, regularise = TRUE)
plotHypercube.regularisation(fitted.cube)
}
