% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypertraps.R
\name{plotHypercube.summary}
\alias{plotHypercube.summary}
\title{Rapidly visualize key metrics from a HyperTraPS run}
\usage{
plotHypercube.summary(
  my.post,
  f.thresh = 0.05,
  t.thresh = 20,
  continuous.time = TRUE
)
}
\arguments{
\item{my.post}{A model fit returned from [HyperTraPS()]}

\item{f.thresh}{Lower threshold passed to [plotHypercube.sampledgraph2()]}

\item{t.thresh}{A upper threshold for time passed to [plotHypercube.timehists()]}

\item{continuous.time}{Whether or not to include [plotHypercube.timehists()]}
}
\value{
a ggplot
}
\description{
Creates a combined likelihood trace, bubble plot and sampledgraph.
}
\examples{
ancestors   <- matrix(c(0,0,0,
                        0,0,1), ncol=3)
descendants <- matrix(c(0,1,1,
                        1,1,1), ncol=3)
start       <- c(0,1)
end         <- c(2,5)
featurenames = c("A", "B", "C")

fitted.cube <- HyperTraPS(descendants, 
                          initialstates = ancestors,
                          starttimes = start,
                          endtimes = end,
                          featurenames = featurenames)
plotHypercube.summary(fitted.cube)
}
