% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypertraps.R
\name{plotHypercube.timehists}
\alias{plotHypercube.timehists}
\title{Timing histograms}
\usage{
plotHypercube.timehists(
  my.post,
  t.thresh = 20,
  featurenames = TRUE,
  log.time = TRUE
)
}
\arguments{
\item{my.post}{A model fit returned from HyperTraPS}

\item{t.thresh}{Upper threshold of time to display. Defaults to 20.}

\item{featurenames}{Either logical (TRUE = use names from model fit), or a character vector of feature names. Default TRUE.}

\item{log.time}{Whether or not to log-transform time. Defaults to TRUE.}
}
\value{
a ggplot
}
\description{
Creates a plot of histograms for timing of each feature's acquisition.
}
\examples{
observations <- matrix(c(0,0,0,
                         0,0,1,
                         0,1,1,
                         1,1,1), byrow=TRUE, ncol=3)
fitted.cube <- HyperTraPS(observations)
plotHypercube.timehists(fitted.cube)
}
