% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypertraps.R
\name{plotHypercube.timeseries}
\alias{plotHypercube.timeseries}
\title{Time series plot}
\usage{
plotHypercube.timeseries(my.post, log.time = TRUE, featurenames = TRUE)
}
\arguments{
\item{my.post}{A model fit returned from HyperTraPS.}

\item{log.time}{Whether or not to log-transform timings. Defaults to TRUE}

\item{featurenames}{Either logical (TRUE = use names from model fit), or a character vector of feature names. Default TRUE.}
}
\value{
a ggplot
}
\description{
Visualize the acquisition of features in continuous time.
}
\examples{
observations <- matrix(c(0,0,0,
                         0,0,1,
                         0,1,1,
                         1,1,1), byrow=TRUE, ncol=3)
fitted.cube <- HyperTraPS(observations)
plotHypercube.timeseries(fitted.cube)
}
