% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PosteriorAnalysis}
\alias{PosteriorAnalysis}
\title{Extracts information from HyperTraPS-related posterior samples}
\usage{
PosteriorAnalysis(
  L,
  featurenames = NULL,
  startstate = NULL,
  use_regularised = 0L,
  limited_output = 0L,
  samples_per_row = 10L,
  outputtransitions = 0L
)
}
\arguments{
\item{L}{A fitted model from HyperTraPS: a named list containing posterior samples}

\item{featurenames}{A character vector containing the set of feature names}

\item{startstate}{A numerical vector describing a particular state from which predictions will be computed}

\item{use_regularised}{Whether to use the regularised (by parameter pruning) version of the parameterisation (1) or not (0). Default 0.}

\item{limited_output}{Whether to limit the terminal output from this process (1) or not (0). Default 0.}

\item{samples_per_row}{How many simulations to run for each posterior sample, in order to characterise that sample behaviour. Default 10.}

\item{outputtransitions}{Whether to output probabilities for each transition in the model space (1) or not (0). Default is 1, but will be switched off for L > 15 because the model space is huge. (Transitions can always be reconstructed from posterior parameter samples)}
}
\value{
Named list containing summary data for feature acquisition ordering ("bubbles"), time histograms, sampled accumulation routes, and timings of these sampled routes.
}
\description{
Extracts information from HyperTraPS-related posterior samples
}
