% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypertraps.R
\name{retention.index}
\alias{retention.index}
\title{Compute retention indices from a set of transitions}
\usage{
retention.index(ct, losses = FALSE, full.output = FALSE)
}
\arguments{
\item{ct}{combined phylogeny and feature data created with [curate.tree()]}

\item{losses}{logical, whether we are considering feature losses (as opposed to gains). Default FALSE}

\item{full.output}{logical, whether to provide full output from the process. Default FALSE}
}
\value{
a dataframe with computed mean and s.d. for retention indices for each feature
}
\description{
The retention index of a feature is the average number of other features retained
when it is lost (in a picture of evolutionary losses), or the average number of
other features absent when it is gained (in a picture of evolutionary gains)
}
