% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypertraps.R
\name{state.probs}
\alias{state.probs}
\title{Get state probabilities from a fitted model}
\usage{
state.probs(my.post, prob.set = NA)
}
\arguments{
\item{my.post}{something}

\item{prob.set}{something}
}
\value{
something
}
\description{
Using a fitted model, return a dataframe with observation probabilities of each state, conditional on a given number of features having been acquired
}
