% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getLikelihood}
\alias{getLikelihood}
\title{Get likelihood for an observation, given a model parameterisation}
\usage{
getLikelihood(
  obs,
  params,
  model,
  initialstate = NULL,
  starttime = NULL,
  endtime = NULL
)
}
\arguments{
\item{obs}{a numeric vector giving the observation}

\item{params}{a numeric vector giving the parameterisation}

\item{model}{the model to use (-1, 1, 2, 3, 4)}

\item{initialstate}{optional: initial state for a transition to the "obs" state (default 0^L)}

\item{starttime}{optional: starting time for the time window for this transition}

\item{endtime}{optional: end time for the time window for this transition}
}
\value{
the likelihood associated with this observation
}
\description{
Get likelihood for an observation, given a model parameterisation
}
