% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypertraps.R
\name{plotHypercube.bubbles.compare}
\alias{plotHypercube.bubbles.compare}
\title{Comparison bubble plot}
\usage{
plotHypercube.bubbles.compare(
  my.post.list,
  reorder = FALSE,
  transpose = FALSE,
  thetastep = 10,
  p.scale = 1,
  sqrt.trans = FALSE,
  bins = 0
)
}
\arguments{
\item{my.post.list}{A list of fitted hypercubes returned from HyperTraPS}

\item{reorder}{Whether to order features by mean acquisition; default FALSE}

\item{transpose}{Whether to transpose the plot; default FALSE}

\item{thetastep}{The number of discrete steps in a polygon approximating a circle segment; default 10 (higher gives smoother visuals but takes longer)}

\item{p.scale}{A scaling factor relating probability to bubble radius; default 1}

\item{sqrt.trans}{Whether to sqrt-transform the probabilities (giving more uniform bubble sizes); default FALSE}

\item{bins}{The number of bins to group time-ordering into; default 0 (do not bin)}
}
\value{
a ggplot
}
\description{
Instead of a "bubble" plot, make bubbles out of a collection of segments, one for each different inference run, so the (dis)agreement between runs can be visualised
}
\examples{
observations <- matrix(c(0,0,0,
                         0,0,1,
                         0,1,1,
                         1,1,1), byrow=TRUE, ncol=3)
fitted.cube.1 <- HyperTraPS(observations, seed=1)
fitted.cube.2 <- HyperTraPS(observations, seed=2)
plotHypercube.bubbles.compare(list(fitted.cube.1, fitted.cube.2))
}
